Update optparse upper bound
I tested this, and everything compiles and appears to parse options correctly.
This allows elm-make to work with base-4.8, which should make building with other packages a lot easier. In particular, I think this might help with getting elm built with GHCJS.
Also, this should resolve the issues with Stackage that arose a while back.